Added functionality for wpm and netwpm
Program now calculates and displays the Gross WPM and Net WPM as the
the user types as well as the current time spent in seconds.

Modifies program to load english and swedish texts separately inorder
to fasten option and texts loads since these arrays are smaller than
the whole text array.